% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gpanel_ellipse.R
\name{gpanel.ellipse}
\alias{gpanel.ellipse}
\title{Lattice panel function for data ellipse}
\usage{
gpanel.ellipse(x, y, ..., radius = 1, n = 100, type)
}
\arguments{
\item{x}{}

\item{y}{}

\item{...}{}
}
\value{
called for its side effects of plotting
}
\description{
Draws a data ellipse in panels
}
\examples{
\dontrun{
library(spida)
library(lattice)
library(latticeExtra)
# No groups or panels:
xyplot(mathach ~ ses, hs) + layer(gpanel.ellipse(...))
# Panels:
xyplot(mathach ~ ses | school, hs) + layer(gpanel.ellipse(...))
# Groups:
xyplot(mathach ~ ses, hs, groups = Sex) + glayer(gpanel.ellipse(...))
# Panels with groups:
xyplot(mathach ~ ses | school, hs, groups = Sex) + glayer(gpanel.ellipse(...))
}
}

