% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fun.R
\name{\%and\%}
\alias{\%and\%}
\alias{\%less\%}
\alias{\%or\%}
\title{Set operators}
\usage{
a \%and\% b
}
\arguments{
\item{a,b}{vectors treated as sets}
}
\description{
Set operations written as binary operators
}

