% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labs.R
\name{labs<-.data.frame}
\alias{labs<-.data.frame}
\title{Add labels -- assignment to data frame}
\usage{
\method{labs}{data.frame}(x) <- value
}
\arguments{
\item{x}{}

\item{value}{}
}
\description{
Add labels -- assignment to data frame
}

