% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spida.R
\name{spida}
\alias{spida}
\title{A package with miscellaneous utilities for hierarchical and longitudinal
data analysis}
\usage{
spida()
}
\description{
This package was developed for the Summer Program in Data Analysis held at
York University from 2000 to 2012. For an overview see
\link{list("spida-package")}.
}

